# map-business-backend/tests/contracts/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "BusinessAgentConfig": {
        "properties": {
          "agent_code": {
            "title": "Agent Code",
            "type": "string"
          },
          "agent_type": {
            "default": "business",
            "title": "Agent Type",
            "type": "string"
          },
          "allowed_roles": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Roles",
            "type": "array"
          },
          "data_scope": {
            "default": "team",
            "title": "Data Scope",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "glossary_terms": {
            "items": {
              "type": "string"
            },
            "title": "Glossary Terms",
            "type": "array"
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "model": {
            "default": "deepseek-v4-flash",
            "title": "Model",
            "type": "string"
          },
          "mounted_resources": {
            "items": {
              "$ref": "#/components/schemas/AgentMountedResource"
            },
            "title": "Mounted Resources",
            "type": "array"
          },
          "owner_team": {
            "title": "Owner Team",
            "type": "string"
          },
          "parallel_limit": {
            "default": 3,
            "title": "Parallel Limit",
            "type": "integer"
          },
          "prompt_config": {
            "$ref": "#/components/schemas/AgentPromptConfig"
          },
          "prompt_template": {
            "default": "",
            "title": "Prompt Template",
            "type": "string"
          },
          "resource_mounts": {
            "items": {
              "$ref": "#/components/schemas/AgentResourceMount"
            },
            "title": "Resource Mounts",
            "type": "array"
          },
          "retry_limit": {
            "default": 1,
            "title": "Retry Limit",
            "type": "integer"
          },
          "scene_name": {
            "title": "Scene Name",
            "type": "string"
          },
          "test_config": {
            "$ref": "#/components/schemas/AgentTestConfig"
          },
          "timeout_s": {
            "default": 120,
            "title": "Timeout S",
            "type": "integer"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "title": "Tools",
            "type": "array"
          },
          "weight": {
            "default": 100,
            "title": "Weight",
            "type": "integer"
          }
        },
        "required": [
          "agent_code",
          "display_name",
          "scene_name",
          "owner_team"
        ],
        "title": "BusinessAgentConfig",
        "type": "object"
      },
      "BusinessAgentTestChatRequest": {
        "properties": {
          "agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BusinessAgentConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "BusinessAgentTestChatRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UploadedSkill": {
        "properties": {
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "mount_agents": {
            "items": {
              "type": "string"
            },
            "title": "Mount Agents",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "skill_id": {
            "title": "Skill Id",
            "type": "string"
          },
          "source": {
            "default": "manual_upload",
            "title": "Source",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "uploaded_at": {
            "title": "Uploaded At",
            "type": "string"
          },
          "version": {
            "default": "1.0.0",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "skill_id",
          "name",
          "display_name",
          "uploaded_at"
        ],
        "title": "UploadedSkill",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "MAP business management BFF. Frontend talks to this service only.",
    "title": "MAP Business Backend",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/business-agents": {
      "get": {
        "operationId": "get_business_agents_api_admin_business_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BusinessAgentConfig"
                  },
                  "title": "Response Get Business Agents Api Admin Business Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Business Agents"
      },
      "post": {
        "operationId": "post_business_agent_api_admin_business_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessAgentConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessAgentConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Business Agent"
      }
    },
    "/api/admin/business-agents/{agent_code}": {
      "put": {
        "operationId": "put_business_agent_api_admin_business_agents__agent_code__put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_code",
            "required": true,
            "schema": {
              "title": "Agent Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessAgentConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessAgentConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Business Agent"
      }
    },
    "/api/admin/business-agents/{agent_code}/test-chat": {
      "post": {
        "operationId": "test_business_agent_api_admin_business_agents__agent_code__test_chat_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_code",
            "required": true,
            "schema": {
              "title": "Agent Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-request-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessAgentTestChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Test Business Agent Api Admin Business Agents  Agent Code  Test Chat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Business Agent"
      }
    },
...
    "/api/admin/mcp-servers": {
      "get": {
        "operationId": "get_mcp_servers_api_admin_mcp_servers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpServerConfig"
                  },
                  "title": "Response Get Mcp Servers Api Admin Mcp Servers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Mcp Servers"
      },
      "post": {
        "operationId": "post_mcp_server_api_admin_mcp_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Mcp Server"
      },
      "put": {
        "operationId": "put_mcp_servers_api_admin_mcp_servers_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/McpServerConfig"
                },
                "title": "Payload",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpServerConfig"
                  },
                  "title": "Response Put Mcp Servers Api Admin Mcp Servers Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Mcp Servers"
      }
    },
    "/api/admin/mcp-servers/{server_id}/refresh-tools": {
      "post": {
        "operationId": "refresh_mcp_server_tools_api_admin_mcp_servers__server_id__refresh_tools_post",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Mcp Server Tools"
      }
    },
...
    "/api/admin/skills": {
      "get": {
        "operationId": "get_uploaded_skills_api_admin_skills_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UploadedSkill"
                  },
                  "title": "Response Get Uploaded Skills Api Admin Skills Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Uploaded Skills"
      },
      "put": {
        "operationId": "put_uploaded_skills_api_admin_skills_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/UploadedSkill"
                },
                "title": "Payload",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UploadedSkill"
                  },
                  "title": "Response Put Uploaded Skills Api Admin Skills Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Uploaded Skills"
      }
    },
    "/api/admin/skills/upload": {
      "post": {
        "operationId": "upload_skill_api_admin_skills_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedSkill"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Skill"
      }
    },
...
  }
}